use codec in sample type

+++ pkg/sample.go
@@ -11,7 +11,7 @@ const (
 // Sample of audio or video
 type Sample struct {
 	ID             string
-	Type           int
+	Type           byte
 	Timestamp      uint32
 	SequenceNumber uint16
 	Payload        interface{}
